+++ Source/CesiumRuntime/Private/CesiumFeaturesMetadataComponent.cpp
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 CesiumGS, Inc. and Contributors
+// Copyright 2020-2025 CesiumGS, Inc. and Contributors
 
 #include "CesiumFeaturesMetadataComponent.h"
 #include "Cesium3DTileset.h"
@@ -46,358 +46,9 @@ extern UNREALED_API class UEditorEngine* GEditor;
 using namespace EncodedFeaturesMetadata;
 using namespace GenerateMaterialUtility;
 
-namespace {
-void AutoFillPropertyTableDescriptions(
-    TArray<FCesiumPropertyTableDescription>& Descriptions,
-    const FCesiumModelMetadata& ModelMetadata) {
-  const TArray<FCesiumPropertyTable>& propertyTables =
-      UCesiumModelMetadataBlueprintLibrary::GetPropertyTables(ModelMetadata);
-
-  for (const auto& propertyTable : propertyTables) {
-    FString propertyTableName = getNameForPropertyTable(propertyTable);
-
-    FCesiumPropertyTableDescription* pDescription =
-        Descriptions.FindByPredicate(
-            [&name = propertyTableName](
-                const FCesiumPropertyTableDescription& existingPropertyTable) {
-              return existingPropertyTable.Name == name;
-            });
-
-    if (!pDescription) {
-      pDescription = &Descriptions.Emplace_GetRef();
-      pDescription->Name = propertyTableName;
-    }
-
-    const TMap<FString, FCesiumPropertyTableProperty>& properties =
-        UCesiumPropertyTableBlueprintLibrary::GetProperties(propertyTable);
-    for (const auto& propertyIt : properties) {
-      auto pExistingProperty = pDescription->Properties.FindByPredicate(
-          [&propertyName = propertyIt.Key](
-              const FCesiumPropertyTablePropertyDescription& existingProperty) {
-            return existingProperty.Name == propertyName;
-          });
-
-      if (pExistingProperty) {
-        // We have already accounted for this property, but we may need to check
-        // for its offset / scale, since they can differ from the class
-        // property's definition.
-        ECesiumMetadataType type = pExistingProperty->PropertyDetails.Type;
-        switch (type) {
-        case ECesiumMetadataType::Scalar:
-        case ECesiumMetadataType::Vec2:
-        case ECesiumMetadataType::Vec3:
-        case ECesiumMetadataType::Vec4:
-        case ECesiumMetadataType::Mat2:
-        case ECesiumMetadataType::Mat3:
-        case ECesiumMetadataType::Mat4:
-          break;
-        default:
-          continue;
-        }
-
-        FCesiumMetadataValue offset =
-            UCesiumPropertyTablePropertyBlueprintLibrary::GetOffset(
-                propertyIt.Value);
-        pExistingProperty->PropertyDetails.bHasOffset |=
-            !UCesiumMetadataValueBlueprintLibrary::IsEmpty(offset);
-
-        FCesiumMetadataValue scale =
-            UCesiumPropertyTablePropertyBlueprintLibrary::GetOffset(
-                propertyIt.Value);
-        pExistingProperty->PropertyDetails.bHasScale |=
-            !UCesiumMetadataValueBlueprintLibrary::IsEmpty(scale);
-
-        continue;
-      }
-
-      FCesiumPropertyTablePropertyDescription& property =
-          pDescription->Properties.Emplace_GetRef();
-      property.Name = propertyIt.Key;
-
-      const FCesiumMetadataValueType ValueType =
-          UCesiumPropertyTablePropertyBlueprintLibrary::GetValueType(
-              propertyIt.Value);
-      property.PropertyDetails.SetValueType(ValueType);
-      property.PropertyDetails.ArraySize =
-          UCesiumPropertyTablePropertyBlueprintLibrary::GetArraySize(
-              propertyIt.Value);
-      property.PropertyDetails.bIsNormalized =
-          UCesiumPropertyTablePropertyBlueprintLibrary::IsNormalized(
-              propertyIt.Value);
-
-      FCesiumMetadataValue offset =
-          UCesiumPropertyTablePropertyBlueprintLibrary::GetOffset(
-              propertyIt.Value);
-      property.PropertyDetails.bHasOffset =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(offset);
-
-      FCesiumMetadataValue scale =
-          UCesiumPropertyTablePropertyBlueprintLibrary::GetOffset(
-              propertyIt.Value);
-      property.PropertyDetails.bHasScale =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(scale);
-
-      FCesiumMetadataValue noData =
-          UCesiumPropertyTablePropertyBlueprintLibrary::GetNoDataValue(
-              propertyIt.Value);
-      property.PropertyDetails.bHasNoDataValue =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(noData);
-
-      FCesiumMetadataValue defaultValue =
-          UCesiumPropertyTablePropertyBlueprintLibrary::GetDefaultValue(
-              propertyIt.Value);
-      property.PropertyDetails.bHasDefaultValue =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(defaultValue);
-
-      property.EncodingDetails = CesiumMetadataPropertyDetailsToEncodingDetails(
-          property.PropertyDetails);
-    }
-  }
-}
-
-void AutoFillPropertyTextureDescriptions(
-    TArray<FCesiumPropertyTextureDescription>& Descriptions,
-    const FCesiumModelMetadata& ModelMetadata) {
-  const TArray<FCesiumPropertyTexture>& propertyTextures =
-      UCesiumModelMetadataBlueprintLibrary::GetPropertyTextures(ModelMetadata);
-
-  for (const auto& propertyTexture : propertyTextures) {
-    FString propertyTextureName = getNameForPropertyTexture(propertyTexture);
-    FCesiumPropertyTextureDescription* pDescription =
-        Descriptions.FindByPredicate(
-            [&propertyTextureName =
-                 propertyTextureName](const FCesiumPropertyTextureDescription&
-                                          existingPropertyTexture) {
-              return existingPropertyTexture.Name == propertyTextureName;
-            });
-
-    if (!pDescription) {
-      pDescription = &Descriptions.Emplace_GetRef();
-      pDescription->Name = propertyTextureName;
-    }
-
-    const TMap<FString, FCesiumPropertyTextureProperty>& properties =
-        UCesiumPropertyTextureBlueprintLibrary::GetProperties(propertyTexture);
-    for (const auto& propertyIt : properties) {
-      auto pExistingProperty = pDescription->Properties.FindByPredicate(
-          [&propertyName =
-               propertyIt.Key](const FCesiumPropertyTexturePropertyDescription&
-                                   existingProperty) {
-            return propertyName == existingProperty.Name;
-          });
-
-      if (pExistingProperty) {
-        // We have already accounted for this property, but we may need to check
-        // for its offset / scale, since they can differ from the class
-        // property's definition.
-        ECesiumMetadataType type = pExistingProperty->PropertyDetails.Type;
-        switch (type) {
-        case ECesiumMetadataType::Scalar:
-        case ECesiumMetadataType::Vec2:
-        case ECesiumMetadataType::Vec3:
-        case ECesiumMetadataType::Vec4:
-        case ECesiumMetadataType::Mat2:
-        case ECesiumMetadataType::Mat3:
-        case ECesiumMetadataType::Mat4:
-          break;
-        default:
-          continue;
-        }
-
-        FCesiumMetadataValue offset =
-            UCesiumPropertyTexturePropertyBlueprintLibrary::GetOffset(
-                propertyIt.Value);
-        pExistingProperty->PropertyDetails.bHasOffset |=
-            !UCesiumMetadataValueBlueprintLibrary::IsEmpty(offset);
-
-        FCesiumMetadataValue scale =
-            UCesiumPropertyTexturePropertyBlueprintLibrary::GetOffset(
-                propertyIt.Value);
-        pExistingProperty->PropertyDetails.bHasScale |=
-            !UCesiumMetadataValueBlueprintLibrary::IsEmpty(scale);
-
-        continue;
-      }
-
-      FCesiumPropertyTexturePropertyDescription& property =
-          pDescription->Properties.Emplace_GetRef();
-      property.Name = propertyIt.Key;
-
-      const FCesiumMetadataValueType ValueType =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::GetValueType(
-              propertyIt.Value);
-      property.PropertyDetails.SetValueType(ValueType);
-      property.PropertyDetails.ArraySize =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::GetArraySize(
-              propertyIt.Value);
-      property.PropertyDetails.bIsNormalized =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::IsNormalized(
-              propertyIt.Value);
-
-      FCesiumMetadataValue offset =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::GetOffset(
-              propertyIt.Value);
-      property.PropertyDetails.bHasOffset =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(offset);
-
-      FCesiumMetadataValue scale =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::GetOffset(
-              propertyIt.Value);
-      property.PropertyDetails.bHasScale =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(scale);
-
-      FCesiumMetadataValue noData =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::GetNoDataValue(
-              propertyIt.Value);
-      property.PropertyDetails.bHasNoDataValue =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(noData);
-
-      FCesiumMetadataValue defaultValue =
-          UCesiumPropertyTexturePropertyBlueprintLibrary::GetDefaultValue(
-              propertyIt.Value);
-      property.PropertyDetails.bHasDefaultValue =
-          !UCesiumMetadataValueBlueprintLibrary::IsEmpty(defaultValue);
-    }
-  }
-}
-
-void AutoFillFeatureIdSetDescriptions(
-    TArray<FCesiumFeatureIdSetDescription>& Descriptions,
-    const FCesiumPrimitiveFeatures& Features,
-    const FCesiumPrimitiveFeatures* InstanceFeatures,
-    const TArray<FCesiumPropertyTable>& PropertyTables) {
-  TArray<FCesiumFeatureIdSet> featureIDSets =
-      UCesiumPrimitiveFeaturesBlueprintLibrary::GetFeatureIDSets(Features);
-  if (InstanceFeatures) {
-    featureIDSets.Append(
-        UCesiumPrimitiveFeaturesBlueprintLibrary::GetFeatureIDSets(
-            *InstanceFeatures));
-  }
-  int32 featureIDTextureCounter = 0;
-
-  for (const FCesiumFeatureIdSet& featureIDSet : featureIDSets) {
-    ECesiumFeatureIdSetType type =
-        UCesiumFeatureIdSetBlueprintLibrary::GetFeatureIDSetType(featureIDSet);
-    int64 count =
-        UCesiumFeatureIdSetBlueprintLibrary::GetFeatureCount(featureIDSet);
-    if (type == ECesiumFeatureIdSetType::None || count == 0) {
-      // Empty or invalid feature ID set. Skip.
-      continue;
-    }
-
-    FString featureIDSetName =
-        getNameForFeatureIDSet(featureIDSet, featureIDTextureCounter);
-    FCesiumFeatureIdSetDescription* pDescription = Descriptions.FindByPredicate(
-        [&name = featureIDSetName](
-            const FCesiumFeatureIdSetDescription& existingFeatureIDSet) {
-          return existingFeatureIDSet.Name == name;
-        });
-
-    if (pDescription) {
-      // We have already accounted for a feature ID set of this name; skip.
-      continue;
-    }
-
-    pDescription = &Descriptions.Emplace_GetRef();
-    pDescription->Name = featureIDSetName;
-    pDescription->Type = type;
-
-    const int64 propertyTableIndex =
-        UCesiumFeatureIdSetBlueprintLibrary::GetPropertyTableIndex(
-            featureIDSet);
-    if (propertyTableIndex >= 0 && propertyTableIndex < PropertyTables.Num()) {
-      const FCesiumPropertyTable& propertyTable =
-          PropertyTables[propertyTableIndex];
-      pDescription->PropertyTableName = getNameForPropertyTable(propertyTable);
-    }
-  }
-}
-
-void AutoFillPropertyTextureNames(
-    TSet<FString>& Names,
-    const FCesiumPrimitiveMetadata& PrimitiveMetadata,
-    const TArray<FCesiumPropertyTexture>& PropertyTextures) {
-  const TArray<int64> propertyTextureIndices =
-      UCesiumPrimitiveMetadataBlueprintLibrary::GetPropertyTextureIndices(
-          PrimitiveMetadata);
-
-  for (const int64& propertyTextureIndex : propertyTextureIndices) {
-    if (propertyTextureIndex < 0 ||
-        propertyTextureIndex >= PropertyTextures.Num()) {
-      continue;
-    }
-
-    const FCesiumPropertyTexture& propertyTexture =
-        PropertyTextures[propertyTextureIndex];
-    FString propertyTextureName = getNameForPropertyTexture(propertyTexture);
-    Names.Emplace(propertyTextureName);
-  }
-}
-
-} // namespace
-
-void UCesiumFeaturesMetadataComponent::AutoFill() {
-  const ACesium3DTileset* pOwner = this->GetOwner<ACesium3DTileset>();
-  if (!pOwner) {
-    return;
-  }
-
-  Super::PreEditChange(NULL);
-
-  // This assumes that the property tables are the same across all models in the
-  // tileset, and that they all have the same schema.
-  for (const UActorComponent* pComponent : pOwner->GetComponents()) {
-    const UCesiumGltfComponent* pGltf = Cast<UCesiumGltfComponent>(pComponent);
-    if (!pGltf) {
-      continue;
-    }
-
-    const FCesiumModelMetadata& modelMetadata = pGltf->Metadata;
-    AutoFillPropertyTableDescriptions(
-        this->Description.ModelMetadata.PropertyTables,
-        modelMetadata);
-    AutoFillPropertyTextureDescriptions(
-        this->Description.ModelMetadata.PropertyTextures,
-        modelMetadata);
-
-    TArray<USceneComponent*> childComponents;
-    pGltf->GetChildrenComponents(false, childComponents);
-
-    for (const USceneComponent* pChildComponent : childComponents) {
-      const auto* pCesiumPrimitive = Cast<ICesiumPrimitive>(pChildComponent);
-      if (!pCesiumPrimitive) {
-        continue;
-      }
-      const CesiumPrimitiveData& primData =
-          pCesiumPrimitive->getPrimitiveData();
-      const FCesiumPrimitiveFeatures& primitiveFeatures = primData.Features;
-      const TArray<FCesiumPropertyTable>& propertyTables =
-          UCesiumModelMetadataBlueprintLibrary::GetPropertyTables(
-              modelMetadata);
-      const FCesiumPrimitiveFeatures* pInstanceFeatures = nullptr;
-      const auto* pInstancedComponent =
-          Cast<UCesiumGltfInstancedComponent>(pChildComponent);
-      if (pInstancedComponent) {
-        pInstanceFeatures = pInstancedComponent->pInstanceFeatures.Get();
-      }
-      AutoFillFeatureIdSetDescriptions(
-          this->Description.PrimitiveFeatures.FeatureIdSets,
-          primitiveFeatures,
-          pInstanceFeatures,
-          propertyTables);
-
-      const FCesiumPrimitiveMetadata& primitiveMetadata = primData.Metadata;
-      const TArray<FCesiumPropertyTexture>& propertyTextures =
-          UCesiumModelMetadataBlueprintLibrary::GetPropertyTextures(
-              modelMetadata);
-      AutoFillPropertyTextureNames(
-          this->Description.PrimitiveMetadata.PropertyTextureNames,
-          primitiveMetadata,
-          propertyTextures);
-    }
-  }
-
-  Super::PostEditChange();
+void UCesiumFeaturesMetadataComponent::AddProperties() {
+  ACesium3DTileset* pOwner = this->GetOwner<ACesium3DTileset>();
+  OnCesiumFeaturesMetadataAddProperties.Broadcast(pOwner);
 }
 
 static FORCEINLINE UMaterialFunction* LoadMaterialFunction(const FName& Path) {
@@ -2034,7 +1685,8 @@ void GenerateNodesForPropertyTexture(
         PropertyTextureName,
         PropertyName);
     ECesiumEncodedMetadataType Type =
-        CesiumMetadataPropertyDetailsToEncodingDetails(Property.PropertyDetails)
+        FCesiumMetadataEncodingDetails::GetBestFitForProperty(
+            Property.PropertyDetails)
             .Type;
 
     if (!foundFirstProperty) {
@@ -2669,24 +2321,24 @@ void UCesiumFeaturesMetadataComponent::PostLoad() {
   // These deprecated variables should only be non-empty on the first load, in
   // which case the contents of Description should be empty.
   if (this->FeatureIdSets.Num() > 0) {
-    CESIUM_ASSERT(this->Description.PrimitiveFeatures.FeatureIdSets.Num == 0);
+    CESIUM_ASSERT(this->Description.PrimitiveFeatures.FeatureIdSets.Num() == 0);
     Swap(
         this->FeatureIdSets,
         this->Description.PrimitiveFeatures.FeatureIdSets);
   }
   if (this->PropertyTextureNames.Num() > 0) {
     CESIUM_ASSERT(
-        this->Description.PrimitiveMetadata.PropertyTextureNames.Num == 0);
+        this->Description.PrimitiveMetadata.PropertyTextureNames.Num() == 0);
     Swap(
         this->PropertyTextureNames,
         this->Description.PrimitiveMetadata.PropertyTextureNames);
   }
   if (this->PropertyTables.Num() > 0) {
-    CESIUM_ASSERT(this->Description.ModelMetadata.PropertyTables.Num == 0);
+    CESIUM_ASSERT(this->Description.ModelMetadata.PropertyTables.Num() == 0);
     Swap(this->PropertyTables, this->Description.ModelMetadata.PropertyTables);
   }
   if (this->PropertyTextures.Num() > 0) {
-    CESIUM_ASSERT(this->Description.ModelMetadata.PropertyTextures.Num == 0);
+    CESIUM_ASSERT(this->Description.ModelMetadata.PropertyTextures.Num() == 0);
     Swap(
         this->PropertyTextures,
         this->Description.ModelMetadata.PropertyTextures);